Move migration hook to global hooks
